+++ examples/master-workers/src/worker.rs
@@ -21,7 +21,7 @@ use crate::task::*;
 
 #[derive(Serialize)]
 pub struct WorkerRegister {
-    pub(crate) speed: u64,
+    pub(crate) speed: f64,
     pub(crate) cpus_total: u32,
     pub(crate) memory_total: u64,
 }
